server/pyproject.toml
[tool.poetry]
name = "url_imageview"
version = "0.1.0"
description = ""
authors = ["patarapolw <patarapolw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
bs4 = "^0.0.1"
pillow = "^5.4"
requests = "^2.21"
flask = "^1.0"
flask-cors = "^3.0"
python-dotenv = "^0.10.1"
pymongo = "^3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
